BG-3970: bugfix

+++ yupana-api/src/main/scala/org/yupana/api/query/Expression.scala
@@ -103,8 +103,8 @@ final case class DistinctCountExpr[I](override val expr: Expression[I])
 }
 
 final case class HLLCountExpr[I](override val expr: Expression[I], accuracy: Double)
-    extends AggregateExpr[I, Set[I], Int](expr, "hll_count") {
-  override val dataType: DataType.Aux[Int] = DataType[Int]
+    extends AggregateExpr[I, Set[I], Long](expr, "hll_count") {
+  override val dataType: DataType.Aux[Long] = DataType[Long]
 }
 
 final case class DistinctRandomExpr[I](override val expr: Expression[I])
